requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langchain-huggingface==0.0.3
sentence-transformers==3.0.1
chromadb>=0.4.24
openai>=1.0.0
anthropic>=0.25.0
python-dotenv>=1.0.1
python-docx==1.1.2
pydantic==2.8.2
click==8.1.7
tqdm==4.66.5
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0

